Model/Model Hub/Torch_ISIC_Inception/requirements.txt
numpy==1.22.3
Pillow==9.1.1
torch==1.11.0
torchvision==0.12.0